% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfft.R
\name{acfft}
\alias{acfft}
\title{Spatial autocorrelation diagnostics}
\usage{
acfft(x, y, f, lim1 = 1, lim2 = 2, dmax = 10)
}
\arguments{
\item{x}{a vector of length n representing the x coordinates
(integer, consecutively numbered cells).}

\item{y}{a vector of length n representing the y coordinates
(integer, consecutively numbered cells)}

\item{f}{residuals from independent regression}

\item{lim1}{lower bound for first bin. Default is 1}

\item{lim2}{upper bound for first bin. Default is 2}

\item{dmax}{number of distance bins to examine. Bins are assumed to be
of uniform size. Default is 10.}
}
\value{
A vector of Moran’s I values for each distance bin.
}
\description{
A function for calculating spatial autocorrelation (i.e. Moran’s I values).
}
