% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.actuals.R
\name{adjusted.actuals}
\alias{adjusted.actuals}
\title{Adjusted actual values}
\usage{
adjusted.actuals(data, coord, plot.maps = FALSE)
}
\arguments{
\item{data}{a dataframe or matrix containing actual presence/absence (binary,
0 or 1) values in 1st column and predicted values (numeric between 0 and 1)
in 2nd column.}

\item{coord}{a matrix of two columns of the same length providing integer,
consecutively numbered coordinates for each occurence and prediction in
\code{data}.}

\item{plot.maps}{A logical indicating whether maps should be plotted.
Default is FALSE.}
}
\value{
a vector of adjusted actual values
}
\description{
ASK GUDRUN WHAT THIS ACTUALLY DOES
}
\examples{
data(hook)
data<- hook[,1:2]
coord<- hook[,3:4]
aa<-adjusted.actuals(data,coord,plot.maps=TRUE)

}
\author{
Gudrun Carl
}
