% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carlinadata}
\alias{carlinadata}
\title{Carlina data set}
\format{A data frame with 961 rows and 5 columns
\describe{
  \item{carlina.horrida}{integer - Simulated count data}
  \item{aridity}{numeric - Simulated aridity index values. This variable has high spatial autocorrelation
  values.}
  \item{land.use}{numeric - Simulated land use intensity. This variable has no spatial autocorrelation.}
  \item{x}{integer - x-coordinates for each grid cell}
  \item{y}{integer - y-coordinates for each grid cell}
}}
\usage{
carlinadata
}
\description{
A data frame containing simulated count data for the thistle, Carlina horrida.
}
\keyword{datasets}
