% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE.R
\name{GEE}
\alias{GEE}
\title{GEE (Generalized Estimating Equations).}
\usage{
GEE(formula, family, data, coord, corstr = "fixed", cluster = 3,
  moran.params = list(), plot = FALSE, graph = FALSE, silent = TRUE)
}
\arguments{
\item{formula}{with specified notation according to names in data frame.}

\item{family}{"gaussian", "binomial"(binary) or "poisson".}

\item{data}{data frame.}

\item{coord}{a matrix of two columns with corresponding cartesian
coordinates, which have to be integer (grid cell numbering).}

\item{corstr}{autocorrelation structure: "independence", "fixed",
"exchangeable", "quadratic"  are possible.
"independence" = GLM, i.e. correlation matrix = identity matrix;
"fixed" for best autocorrelation removal by means of an adapted
isotropic power function specifying all correlation
coefficients "exchangeable", "quadratic" for clustering, i.e.
the correlation matrix has a block diagonal form:
"exchangeable": all intra-block correlation coefficients are equal;
"quadratic": intra-block correlation coefficients for different
distances can be different.}

\item{cluster}{cluster size for cluster models: "exchangeable", "quadratic".
cluster= 2,3,4 are allowed for size 2*2,3*3,4*4, respectively.}

\item{plot}{a logical value indicating whether results should be plotted.}

\item{graph}{a logical value indicating whether results should be displayed.}

\item{moran}{a list of parameters for calculating Moran's I.
This is passed to wrm.moran.}
}
\value{
An object of class "gee" is a list containing the following components:
         call       call
         formula    formula
         family     family
         b          estimate of regression parameters
         s.e.       standard errors
         z          z values (or corresponding values for statistics)
         p          probabilities
         scale      scale parameter (dispersion parameter)
         fitted     fitted values
         resid      normalized Pearson residuals
         w.ac       working autocorrelation parameters
         W.ac       working autocorrelation matrix
         QIC        quasi-information criterion
         if plot or graph is true:
         ac.glm     autocorrelation of glm.residuals
         ac         autocorrelation of gee.residuals
}
\description{
\code{GEE} provides a GEE-based method to account for spatial
autocorrelation in multiple linear regressions
}
\details{
GEE can be used to fit linear models for response vectors of  different
distributions: "gaussian", "binomial"(binary) or "poisson". As spatial
model, it is a generalized linear model in which the residuals may be
autocorrelated. It accounts for spatial (2-dimensional) residual
autocorrelation in case of regular gridded datasets. The grid cells
are assumed to be square.
}
\references{
Carl & Kühn (2007): Analyzing Spatial Autocorrelation in Species
Distributions using Gaussian and Logit Models, Ecol. Model. 207, 159 - 170
}
