% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmiWMRR.R
\name{mmiWMRR}
\alias{mmiWMRR}
\title{Multi-model inference for wavelet multiresolution regression}
\usage{
mmiWMRR(formula, family, data, coord, scale, detail = TRUE,
  wavelet = "haar", wtrafo = "dwt", n.eff = NULL)
}
\arguments{
\item{formula}{With specified notation according to names in data frame.}

\item{family}{\code{gaussian}, \code{binomial}, or \code{poisson}.}

\item{data}{Data frame}

\item{coord}{Corresponding coordinates which have to be integer}

\item{scale}{0 or higher integers possible (limit depends on sample size)}

\item{detail}{Remove non-smooth wavelets? Default is \code{TRUE}}

\item{wavelet}{Name of wavelet family. \code{haar}, \code{d4}, and \code{la8}.
are possible. \code{haar} is the default.}

\item{wtrafo}{Type of wavelet transform. Either \code{dwt} or \code{modwt}.
\code{dwt} is the default.}

\item{n.eff}{A numeric value of effective sample size}
}
\value{
\code{mmiWMRR} returns a list containing the following elements
\describe{
  \item{\code{result}}{A matrix containing slopes, degrees of freedom, likelihood,
         AIC, delta, and weight values for the set of candidate models.
         The models are ranked by Akaike weights.}
  \item{\code{level}}{An integer corresponding to scale}

}
}
\description{
mmiWMRR is a multimodel inference approach evaluating the relative
importance of predictors used in \code{\link{scaleWMRR}}. It performs automatically
generated model selection and creates a model
selection table according to the approach of multi-model inference
(Burnham & Anderson, 2002). The analysis is carried out for scale-specific
regressions (i.e. where \code{\link{scaleWMRR}} can be used). AIC is used to obtain model
selection weights and to rank the models.
}
\examples{

data(carlinadata)
coord<- carlinadata[,4:5]

# scale-specific regressions for detail components
# ranked by multimodel inference
A<-array(NA,c(4,8,4))
level<-rep(NA,4)
for (i in 1:4) {
  mmi<- mmiWMRR(carlina.horrida ~ aridity + land.use,"poisson",
                carlinadata,coord,scale=i,detail=TRUE,wavelet="d4")
  A[,,i]<-mmi$result
  level[i]<-mmi$level
}

# Plot scale-dependent relative variable importance
rvi.plot(carlina.horrida ~ aridity + land.use,"poisson",
         carlinadata,coord,maxlevel=4,detail=TRUE,wavelet="d4")

}
\references{
Burnham, K.P. & Anderson, D.R. (2002) Model selection and
multimodel inference. Springer, New York.

Carl G, Doktor D, Schweiger O, Kuhn I (2016)
Assessing relative variable importance across different spatial
scales: a two-dimensional wavelet analysis.
Journal of Biogeography 43: 2502-2512.
}
\seealso{
\code{\link{aic.calc}}, \pkg{MuMIn}.
}
