% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE.R
\name{qic}
\alias{qic}
\title{Quasi-Information Criterion for Generalized Estimating
Equations}
\usage{
qic(formula, data, family, mu, var.robust, var.indep.naive)
}
\arguments{
\item{formula}{with specified notation according to names in data frame}

\item{data}{data frame}

\item{family}{"gaussian", "binomial", or "poisson"}

\item{mu}{fitted values}

\item{var.robust}{variance of b values}

\item{var.indep.naive}{naive variance of b values under the
"independence" model}
}
\value{
A list with the following components:

\code{QIC}        quasi-information criterion

\code{loglik}     quasi-likelihood
}
\description{
A function for calculating quasi-likelihood and Quasi-Information
Criterion values based on the method of Hardin & Hilbe (2003).
}
\references{
QIC as defined in Hardin & Hilbe (2003).#This is confusing, ask for original H&H paper for this section

Barnett et al. Methods in Ecology & Evolution 2010, 1, 1524.
}
