% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvi_plot.R
\name{rvi.plot}
\alias{rvi.plot}
\title{Model selection tables and relative variable importance
in terms of scale levels}
\usage{
rvi.plot(formula, family, data, coord, maxlevel, detail = TRUE,
  wavelet = "haar", wtrafo = "dwt", n.eff = NULL)
}
\arguments{
\item{formula}{A model formula}

\item{family}{Error structure. \code{Gaussian}, \code{binomial}, and \code{Poisson}
are supported.}

\item{data}{A data frame or set of vectors of equal length.}

\item{coord}{X,Y coordinates for each observation. Coordinates should be
consecutive integers.}

\item{maxlevel}{An integer for maximum scale level}

\item{detail}{By detail components only}

\item{wavelet}{Type of wavelet: \code{haar}, \code{d4}, or \code{la8}}

\item{wtrafo}{Type of wavelet transform: \code{dwt} or \code{modwt}}

\item{n.eff}{A numeric value of effective sample size}
}
\description{
The summary is based on the results provided by
multi-model inference for wavelet multiresolution regression (mmiWMRR).
Based on the results of this approach, the relative
importance of predictors are calculated. The scale-dependent results
are graphically displayed.
}
