% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_WMRR.R
\name{scaleWMRR}
\alias{scaleWMRR}
\title{Scaling by wavelet multiresolution regression (WMRR)}
\usage{
scaleWMRR(formula, family, data, coord, scale = 1, detail = TRUE,
  wavelet = "haar", wtrafo = "dwt", b.ini = NULL, pad = list(),
  control = list(), moran = list(), plot = FALSE)
}
\arguments{
\item{formula}{with specified notation according to names in data frame}

\item{family}{"gaussian", "binomial"(binary) or "poisson"}

\item{data}{data frame}

\item{coord}{corresponding coordinates which have to be integer}

\item{scale}{0 (which is equivalent to GLM) or
higher integers possible (limit depends on sample size)}

\item{detail}{by detail components only,
i.e. smooth wavelet components are removed in this case}

\item{wavelet}{type of wavelet: "haar" or "d4" or "la8"}

\item{wtrafo}{type of wavelet transform: "dwt" or "modwt"}
}
\value{
scaleWMRR returns a list containing the following elements
\describe{
  \item{\code{call}}{call}
  \item{\code{b}}{estimate of regression parameters}
  \item{\code{s.e.}}{standard errors}
  \item{\code{z}}{z values (or corresponding values for statistics)}
  \item{\code{p}}{probabilities}
  \item{\code{df}}{degrees of freedom}
  \item{\code{fitted}}{fitted values}
  \item{\code{resid}}{Pearson residuals}
  \item{\code{converged}}{logical value whether the procedure converged}
  \item{if plot is true}
    \itemize{
      \item{\code{ac.glm}}{autocorrelation of glm.residuals}
      \item{\code{ac}}{autocorrelation of wavelet.residuals}
  }

}
}
\description{
scaleWMRR performs a scale-specific regression based on a
wavelet multiresolution analysis. It can be used to fit
generalized linear models taking the two-dimensional grid structure of
datasets into account. The following error distributions (in
conjunction with appropriate link functions) are allowed: "gaussian",
"binomial"(binary) or "poisson". The model provides scale-specific
results for data sampled on a contiguous geographical area. The
dataset is assumed to be regular gridded and the grid cells are
assumed to be square. For the wavelet transforms, a function
of package 'waveslim' is used (Whitcher, 2005).
}
\references{
Carl G, Doktor D, Schweiger O, K?hn I (2016)
Assessing relative variable importance across different spatial
scales: a two-dimensional wavelet analysis.
Journal of Biogeography 43: 2502?2512.

Whitcher, B. (2005) Waveslim: basic wavelet routines for one-, two-
and three-dimensional signal processing. R package version 1.5.
}
