% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spind.R
\docType{package}
\name{spind}
\alias{spind}
\alias{spind-package}
\title{spind: a package for computing spatially corrected accuracy measures}
\description{
The \code{spind} package provides convenient implementation of Generalized
estimating equations (GEEs) and Wavelet-revised models (WRMs)
in the context of spatial models. It also provides tools for
multi-model inference, stepwise model selection, and spatially
corrected model diagnostics. This help section provides brief descriptions
of each function and is organized by the type of model they apply to
or the scenarios in which you might use them. Of course, these are
recommendations - feel free to use them as you see fit. For a more
detailed description of the package and its functions, please see
the vignette \emph{Intro to spind} (\code{browseVignettes('spind')}).
}
