% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_GEE.R
\name{summary.GEE}
\alias{summary.GEE}
\title{Summarize the output from\code{GEE}}
\usage{
\method{summary}{GEE}(object, ..., printAutoCorPars = TRUE)
}
\arguments{
\item{object}{An object of class \code{GEE}}

\item{...}{Additional parameters to be passed summary.}

\item{printAutoCorPars}{A logical indicating whether to print the
autocorrelation matrix}
}
\value{
Prints model details, parameter estimates, and autocorrelation values
for the first 10 distance bins. Additionally, if \code{printAutoCorPars} = TRUE,
prints autocorrelation parameters used in the model.
}
\description{
Returns summary of GEE parameter estimates and autocorrelations
of the residuals.
}
\author{
Sam Levin
}
