% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_WRM.R
\name{summary.WRM}
\alias{summary.WRM}
\title{Summarize the output from \code{WRM}}
\usage{
\method{summary}{WRM}(object, ...)
}
\arguments{
\item{object}{A model of class \code{WRM}}

\item{...}{Other arguments passed to \code{summary}}
}
\value{
Prints the results of a WRM
}
\description{
Returns summary of WRM parameter estimates and autocorrelations
of the residuals.
}
\author{
Gudrun Carl, Sam Levin
}
