% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_dep.R
\name{th.dep}
\alias{th.dep}
\title{th.dep}
\usage{
th.dep(data, coord, thresh = 0.5, spatial = TRUE)
}
\arguments{
\item{data}{A data frame or matrix with two columns. The first column
should contain actual presence/absence data (binary, 0 or 1) and the
second column should contain model predictions of probability of
occurence (numeric, between 0 and 1).}

\item{coord}{A data frame or matrix with two columns containing x,y
coordinates for each actual and predicted value. Coordinates must be
integer and consecutively numbered.}

\item{thresh}{A cutoff value for classifying predictions as modeled
presences or modeled absences. Default is 0.5.}

\item{spatial}{A logical indcating whether spatially corrected indices
(rather than classical indices) should be computed.}
}
\value{
A list with the following components
\describe{
 \item{kappa}{kappa}
 \item{cm}{confusion matrix}
 \item{sensitivity}{sensitivity}
 \item{specificity}{specificity}
 \item{actuals}{actuals or adjusted actuals}
 \item{splitlevel.pred}{level splitting of predicted values}
 \item{splitlevel.act}{level splitting of actuals or adjusted actuals}
 \item{splitposition.pred}{position splitting of predicted values}
 \item{splitposition.act}{position splitting of actuals or adjusted actuals}
}
}
\description{
Calculates threshold-dependent metrics for an observational
data set and model predictions. It is a 2-dimensional analysis that takes
the grid structure of the data into account. Grid cells are assumed to be
square
}
\details{
Kappa and confusion matrix for spatial models
}
\examples{
data(hook)
data<-hook[,1:2]
coord<-hook[,3:4]
si1<-th.dep(data,coord,spatial=TRUE)
si1$kappa
si1$cm

}
\author{
Gudrun Carl
}
