% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavecovar.R
\name{wavecovar}
\alias{wavecovar}
\title{Wavelet covariance analysis}
\usage{
wavecovar(f1, f2, x, y, wavelet = "haar", wtrafo = "dwt")
}
\arguments{
\item{f1}{a vectors}

\item{f2}{a vectors of same length}

\item{x}{corresponding x-coordinates which have to be integer}

\item{y}{corresponding y-coordinates which have to be integer}

\item{wavelet}{type of wavelet: "haar" or "d4" or "la8"}

\item{wtrafo}{type of wavelet transform: "dwt" or "modwt"}
}
\value{
Wavelet covariance
}
\description{
The analysis is based a wavelet multiresolution analysis.
It is a 2D analysis taking the grid structure of datasets into
account, i.e. the analysis provides scale-specific
results for data sampled on a contiguous geographical area. The
dataset is assumed to be regular gridded and the grid cells are
assumed to be square.
}
\seealso{
\pkg{waveslim}, \code{\link{WRM}}
}
