% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRM.R
\name{WRM}
\alias{WRM}
\title{Wavelet-revised models (WRMs)}
\usage{
WRM(formula, family, data, coord, level = 1, wavelet = "haar",
  wtrafo = "dwt", b.ini = NULL, pad = list(), control = list(),
  moran.params = list(), plot = FALSE, graph = FALSE)
}
\arguments{
\item{formula}{Model formula. Variable names must match variables in \code{data}.}

\item{family}{\code{gaussian}, \code{binomial}, or \code{poisson} are supported.}

\item{data}{A data frame with variable names that match the variables specified in \code{formula}.}

\item{coord}{A matrix of two columns with corresponding cartesian
coordinates. Currently only supports integer coordinates.}

\item{level}{An integer specifying the depth of wavelet decomposition
\itemize{
     \item{0} Without autocorrelation removal (equivalent to a GLM)
     \item{1} For best autocorrelation removal
     \item{...} Higher integers possible. The limit depends on sample size # We should expand on this. How does one determin the limit?
  }}

\item{wavelet}{Name of wavelet family. \code{haar}, \code{d4}, and \code{la8}
are possible.}

\item{wtrafo}{Type of wavelet transform. Either \code{dwt} or \code{modwt} # briefly explain differences?}

\item{b.ini}{Initial parameter estimates. Default is NULL.}

\item{pad}{A list of parameters for padding wavelet coefficients.
\itemize{
   \item{padform} 0, 1, and 2 are possible.
    \code{padform} is automatically set to
    zero when either \code{level}=0 or
    a \code{formula} including an intercept and a non-gaussian family
   \itemize{
     \item{0} Padding with 0s.
     \item{1} Padding with mean values.
     \item{2} Padding with mirror values.
 }
   \item{padzone} Factor for expanding the padding zone
}}

\item{control}{a list of parameters for controlling the fitting process.
   \itemize{
      \item{eps} Positive convergence tolerance
      \item{denom.eps} The iterations converge when
           \eqn{max(|coeff.new-coeff.old|)/(|coeff.old|+denom.eps) <= eps}
      \item{itmax} Integer giving the maximum number of iterations
}}

\item{moran.params}{A list of parameters for calculating Moran's I.
\itemize{
  \item\code{lim1} Lower limit for first bin. Default is 0.
  \item\code{increment} Step size for calculating I. Default is 1.
}}

\item{plot}{a logical value indicating whether results should be plotted}

\item{graph}{a logical value indicating whether results should be displayed}
}
\value{
An object of class "wrm" is a list containing the following components:
         call       call
         formula    formula
         family     family
         b          estimate of regression parameters
         s.e.       standard errors
         z          z values (or corresponding values for statistic)
         p          probabilities
         fitted     fitted values
         resid      Pearson residuals
         b.sm       estimate of regression parameters of neglected smooth part
         fitted.sm  fitted values of neglected smooth part
         level      level
         wavelet    wavelet
         wtrafo     wtrafo
         padzone    padzone
         padform    padform
         n.eff      effective number of observations
         AIC        Akaike information criterion
         if plot or graph is true:
         ac.glm     autocorrelation of glm.residuals
         ac         autocorrelation of wavelet.residuals
}
\description{
A wavelet-based method to remove spatial autocorrelation
in multiple linear regressions.
}
\details{
WRM can be used to fit linear models for response vectors of different
distributions: \code{gaussian}, \code{binomial}, or \code{poisson}.
As a spatial model, it is a generalized linear model in which the residuals
may be autocorrelated. It corrects for  2-dimensional residual
autocorrelation for regular gridded data sets using the wavelet
decomposition technique. The grid cells are assumed to be square.
}
\references{
Carl, G., Kuhn, I. (2010): A wavelet-based extension of generalized
linear models to remove the effect of spatial autocorrelation.
Geographical Analysis 42 (3), 323 - 337
}
\author{
Gudrun Carl
}
