% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covar.R
\name{covar.plot}
\alias{covar.plot}
\title{Plot wavelet variance/covariance}
\usage{
covar.plot(formula, data, coord, wavelet = "haar", wtrafo = "dwt",
  plot = "covar")
}
\arguments{
\item{formula}{With specified notation according to names in data frame}

\item{data}{Data frame}

\item{coord}{A matrix of 2 columns with
corresponding x,y-coordinates which have to be integer}

\item{wavelet}{Type of wavelet: \code{haar}, \code{d4}, or \code{la8}}

\item{wtrafo}{Type of wavelet transform: \code{dwt} or \code{modwt}}

\item{plot}{Either \code{var} for wavelet variance analysis
or \code{covar} for wavelet covariance analysis}
}
\value{
A list containing a vector of results
}
\description{
The plot is based on a wavelet variance or covariance analysis
and will work where \code{\link{wavevar}} or \code{\link{wavecovar}} as function
is supported. The scale-dependent results are graphically displayed.
}
