% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE.R
\name{GEE}
\alias{GEE}
\title{GEE (Generalized Estimating Equations)}
\usage{
GEE(formula, family, data, coord, corstr = "fixed", cluster = 3,
  moran.params = list(), plot = FALSE)
}
\arguments{
\item{formula}{Model formula. Variable names must match variables in \code{data}.}

\item{family}{\code{gaussian}, \code{binomial}, or \code{poisson} are supported.}

\item{data}{A data frame with variable names that match the variables specified in \code{formula}.}

\item{coord}{A matrix of two columns with corresponding cartesian
coordinates. Currently only supports integer coordinates.}

\item{corstr}{Expected autocorrelation structure: \code{independence}, \code{fixed},
\code{exchangeable}, and \code{quadratic}  are possible.

 \itemize{
   \item\code{independence} is the same as a GLM, i.e. correlation matrix = identity matrix;

   \item\code{fixed} for best autocorrelation removal by means of an adapted
   isotropic power function specifying all correlation
   coefficients

   \item\code{exchangeable} and \code{quadratic} for clustering, i.e.
   the correlation matrix has a block diagonal form:

   \itemize{
      \item\code{exchangeable}: all intra-block correlation coefficients are equal

      \item\code{quadratic}: intra-block correlation coefficients for different
         distances can be different.
         }
       }}

\item{cluster}{Cluster size for cluster models \code{exchangeable}
 and \code{quadratic}. values of 2, 3, and 4 are allowed
 \itemize{
   \item 2 - a 2*2 cluster

   \item 3 - a 3*3 cluster

   \item 4 - a 4*4 cluster
}}

\item{moran.params}{A list of parameters for calculating Moran's I.
\itemize{
  \item\code{lim1} Lower limit for first bin. Default is 0.
  \item\code{increment} Step size for calculating I. Default is 1.
}}

\item{plot}{A logical value indicating whether autocorrelation of
residuals should be plotted.}
}
\value{
An object of class \code{GEE}. This consists of a list with the
following elements:
\describe{
      \item{\code{call}}{Call}
      \item{\code{formula}}{Model formula}
      \item{\code{family}}{Family}
      \item{\code{corstr}}{User-selected correlaton structure}
      \item{\code{b}}{Estimate of regression parameters}
      \item{\code{s.e.}}{Standard errors of the estimates}
      \item{\code{z}}{Depending on the \code{family}, either a z or t value}
      \item{\code{p}}{p-values}
      \item{\code{scale}}{scale parameter (dispersion parameter)}
      \item{\code{fitted}}{Fitted values}
      \item{\code{resid}}{Normalized Pearson residuals}
      \item{\code{w.ac}}{Working autocorrelation parameters}
      \item{\code{Mat.ac}}{Working autocorrelation matrix}
      \item{\code{QIC}}{Quasi-information criterion. See \code{\link{qic}}
       for further details}
      \item{\code{plot}}{Logical value indicating whether autocorrelation should
      be plotted}
      \item{\code{ac.glm}}{Autocorrelation of GLM residuals}
      \item{\code{ac.gee}}{Autocorrelation of GEE residuals}
}

Elements can be viewed using the \code{\link{summary.GEE}} function included in
the package
}
\description{
\code{GEE} provides a GEE-based method to account for spatial
autocorrelation in multiple linear regressions
}
\details{
GEE can be used to fit linear models for response variables with
different distributions: \code{gaussian}, \code{binomial}, or \code{poisson}.
As a spatial model, it is a generalized linear model in which the residuals
may be autocorrelated. It accounts for spatial (2-dimensional) residual
autocorrelation in cases of regular gridded datasets. The grid cells
are assumed to be square.
}
\examples{
data(musdata)
coords<- musdata[,4:5]

x<-GEE(musculus ~ pollution + exposure, "poisson", musdata,
      coord=coords, corstr="exchangeable", plot=TRUE)

summary(x,printAutoCorPars=TRUE)

}
\references{
Carl & Kühn (2007): Analyzing Spatial Autocorrelation in Species
Distributions using Gaussian and Logit Models, Ecol. Model. 207, 159 - 170
}
\seealso{
\code{\link{qic}}, \code{\link{summary.GEE}}
}
