% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE.R
\name{GEE}
\alias{GEE}
\title{GEE (Generalized Estimating Equations)}
\usage{
GEE(formula, family, data, coord, corstr = "fixed", cluster = 3,
  moran.params = list(), plot = FALSE, silent = TRUE)
}
\arguments{
\item{formula}{variable names must match variables in \code{data}.}

\item{family}{\code{gaussian}, \code{binomial}, or \code{poisson} are supported.}

\item{data}{a data frame with names that match the variables specified in \code{formula}.}

\item{coord}{a matrix of two columns with corresponding cartesian
coordinates, which have to be integer (grid cell numbering).}

\item{corstr}{expected autocorrelation structure: "independence", "fixed",
"exchangeable", "quadratic"  are possible.

 \itemize{
   \item\emph{independence} is the same as a GLM, i.e. correlation matrix = identity matrix;

   \item\emph{fixed} for best autocorrelation removal by means of an adapted
   isotropic power function specifying all correlation
   coefficients

   \item\emph{exchangeable} and \emph{quadratic} for clustering, i.e.
   the correlation matrix has a block diagonal form:

   \itemize{
      \item\emph{exchangeable}: all intra-block correlation coefficients are equal

      \item\emph{quadratic}: intra-block correlation coefficients for different
         distances can be different.
         }
       }}

\item{cluster}{cluster size for cluster models \emph{exchangeable}
 and \emph{quadratic}. values of 2, 3, and 4 are allowed
 \itemize{
   \item 2 - a 2*2 cluster

   \item 3 - a 3*3 cluster

   \item 4 - a 4*4 cluster
 }}

\item{moran.params}{a list of parameters for calculating Moran's I.
These are passed to \code{wrm.moran} internally.}

\item{plot}{a logical value indicating whether results should be plotted.}

\item{silent}{a logical value controlling whether parameter estimates are
printed after each iteration}
}
\value{
An object of class "gee". List components can be viewed using the
\code{summary_gee} function
}
\description{
\code{GEE} provides a GEE-based method to account for spatial
autocorrelation in multiple linear regressions
}
\details{
GEE can be used to fit linear models for response variables with
different distributions: "gaussian", "binomial", or "poisson".
As a spatial model, it is a generalized linear model in which the residuals
may be autocorrelated. It accounts for spatial (2-dimensional) residual
autocorrelation in cases of regular gridded datasets. The grid cells
are assumed to be square.
}
\references{
Carl & Kühn (2007): Analyzing Spatial Autocorrelation in Species
Distributions using Gaussian and Logit Models, Ecol. Model. 207, 159 - 170
}
