% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmiWMRR.R
\name{mmiWMRR}
\alias{mmiWMRR}
\title{Multi-model inference for wavelet multiresolution regression}
\usage{
mmiWMRR(formula, family, data, coord, scale, detail = TRUE,
  wavelet = "haar", wtrafo = "dwt", n.eff = NULL)
}
\arguments{
\item{formula}{with specified notation according to names in data frame}

\item{family}{"gaussian", "binomial"(binary) or "poisson"}

\item{data}{data frame}

\item{coord}{corresponding coordinates which have to be integer}

\item{scale}{0 or higher integers possible (limit depends on sample size)}

\item{detail}{by detail components only}

\item{wavelet}{type of wavelet: "haar" or "d4" or "la8"}

\item{wtrafo}{type of wavelet transform: "dwt" or "modwt"}

\item{n.eff}{a numeric value of effective sample size}
}
\value{
mmiWMRR returns a list containing the following elements
\describe{
  \item{\code{result}}{a matrix containing slopes, degrees of freedom, Likelihood,
         AIC, delta and weight values for the set of candidate models.
         The models are ranked by Akaike weights.}
  \item{\code{level}}{an integer corresponding to scale}

}
}
\description{
mmiWMRR is a multimodel inference approach evaluating the relative
importance of predictors used in scaleWMRR. It performs automatically
generated model selection and creates a model
selection table according to the approach of multi-model inference
(Burnham & Anderson, 2002). The analysis is carried out for scale-specific
regressions, i.e. where scaleWMRR as
modelling function is supported. AIC is used to obtain model
selection weights and to rank the models.
}
\references{
Burnham, K.P. & Anderson, D.R. (2002) Model selection and
multimodel inference. Springer, New York.

Carl G, Doktor D, Schweiger O, Kuhn I (2016)
Assessing relative variable importance across different spatial
scales: a two-dimensional wavelet analysis.
Journal of Biogeography 43: 2502-2512.
}
\seealso{
AIC, \pkg{MuMIn}.
}
