% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{musdata}
\alias{musdata}
\title{Mus musculus data set}
\format{A data frame with 400 rows and 5 columns
 \describe{
   \item{musculus}{integer - Simulated count data for Mus musculus}
   \item{pollution}{numeric - Simulated variable that describes degree
    of pollution in corresponding grid cell}
   \item{exposure}{numeric - Simulated variable that describes degree of
    exposure for each grid cell}
   \item{x}{integer - x-coordinates for each grid cell}
   \item{y}{integer - y-coordinates for each grid cell}
 }}
\usage{
musdata
}
\description{
A data frame containing simulated count data of a house mouse.
}
\keyword{datasets}
