% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{musdata}
\alias{musdata}
\title{Mus musculus data set}
\format{A data frame with 400 rows and 5 columns
 \describe{
   \item{musculus}{integer - Presence/absence records of Mus musculus}
   \item{pollution}{numeric - describes degree of pollution in corresponding grid cell}
   \item{exposure}{numeric - describes degree of exposure for each grid cell}
   \item{x}{integer - x-coordinates for each grid cell}
   \item{y}{integer - y-coordinates for each grid cell}
 }}
\usage{
musdata
}
\description{
A data frame containing simulated presence/absence records of a hypothetical house mouse.
}
\keyword{datasets}
