% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_WMRR.R
\name{scaleWMRR}
\alias{scaleWMRR}
\title{Scaling by wavelet multiresolution regression (WMRR)}
\usage{
scaleWMRR(formula, family, data, coord, scale = 1, detail = TRUE,
  wavelet = "haar", wtrafo = "dwt", b.ini = NULL, pad = list(),
  control = list(), moran.params = list(), plot = FALSE)
}
\arguments{
\item{formula}{with specified notation according to names in data frame}

\item{family}{"gaussian", "binomial"(binary) or "poisson"}

\item{data}{data frame}

\item{coord}{corresponding coordinates which have to be integer}

\item{scale}{0 (which is equivalent to GLM) or
higher integers possible (limit depends on sample size)}

\item{detail}{by detail components only,
i.e. smooth wavelet components are removed in this case}

\item{wavelet}{type of wavelet: "haar" or "d4" or "la8"}

\item{wtrafo}{type of wavelet transform: "dwt" or "modwt"}

\item{b.ini}{Initial parameter estimates. Default is NULL.}

\item{pad}{A list of parameters for padding wavelet coefficients.
\itemize{
   \item{padform} - 0, 1, and 2 are possible.
    \code{padform} is automatically set to
    0 when either \code{level}=0 or
    a \code{formula} including an intercept and a non-gaussian family
   \itemize{
     \item{0} - Padding with 0s.
     \item{1} - Padding with mean values.
     \item{2} - Padding with mirror values.
 }
   \item{padzone} - Factor for expanding the padding zone
}}

\item{control}{a list of parameters for controlling the fitting process.
   \itemize{
      \item{\code{eps}} - Positive convergence tolerance. Smaller values of
      \code{eps} provide better parameter estimates, but also reduce the probability
      of the iterations converging. In case of issues with convergence, test larger
      values of \code{eps}. Default is 10^-5.
      \item{\code{denom.eps}} - Default is 10^-20.
      \item{\code{itmax}} - Integer giving the maximum number of iterations.
      Default is 200.
}}

\item{moran.params}{A list of parameters for calculating Moran's I.
\itemize{
  \item\code{lim1} - Lower limit for first bin. Default is 0.
  \item\code{increment} - Step size for calculating Moran's I. Default is 1.
}}

\item{plot}{A logical value indicating whether to print parameter estimates
to the console}
}
\value{
scaleWMRR returns a list containing the following elements
\describe{
  \item{\code{call}}{call}
  \item{\code{b}}{estimate of regression parameters}
  \item{\code{s.e.}}{standard errors}
  \item{\code{z}}{z values (or corresponding values for statistics)}
  \item{\code{p}}{probabilities}
  \item{\code{df}}{degrees of freedom}
  \item{\code{fitted}}{fitted values}
  \item{\code{resid}}{Pearson residuals}
  \item{\code{converged}}{logical value whether the procedure converged}
  \item{\code{plot}}{Logical. If TRUE:}

    \itemize{
      \item\code{ac.glm} autocorrelation of glm.residuals

      \item\code{ac} autocorrelation of wavelet.residuals
  }

}
}
\description{
scaleWMRR performs a scale-specific regression based on a
wavelet multiresolution analysis. It can be used to fit
generalized linear models taking the two-dimensional grid structure of
datasets into account. The following error distributions (in
conjunction with appropriate link functions) are allowed: "gaussian",
"binomial"(binary) or "poisson". The model provides scale-specific
results for data sampled on a contiguous geographical area. The
dataset is assumed to be regular gridded and the grid cells are
assumed to be square. For the wavelet transforms, a function
of package 'waveslim' is used (Whitcher, 2005).
}
\references{
Carl G, Doktor D, Schweiger O, Kuhn I (2016)
Assessing relative variable importance across different spatial
scales: a two-dimensional wavelet analysis.
Journal of Biogeography 43: 2502?2512.

Whitcher, B. (2005) Waveslim: basic wavelet routines for one-, two-
and three-dimensional signal processing. R package version 1.5.
}
