% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_spind.R
\name{step.spind}
\alias{step.spind}
\title{Stepwise model selection for WRMs and GEEs}
\usage{
step.spind(object, data, scope = NULL, direction = "backward")
}
\arguments{
\item{object}{A model fit of class \code{WRM} or \code{GEE}.}

\item{data}{The data used to fit the specified model.}

\item{scope}{A list with components \code{upper} and \code{lower}
that specify the upper and lower boundaries of the stepwise
model selection procedure. Default is NULL, which will step backwards through
all predictor variables.}

\item{direction}{Only \code{backward} is supported right now. The operation
will begin with the parameter \code{upper} and work backwards until it reaches
parameter \code{lower} (specified in \code{scope}). \code{forward} or \code{both} may be added later.}
}
\value{
A data frame containing model formulae. WRM models will have AIC
Log-likelihood, and delta-AIC information. GEE models will have QIC,
Quasi-likelihood, and delta-QIC information.
}
\description{
A function that provides stepwise model selection using
AIC for Wavelet revised models (WRMs) and QIC for Generalized estimating
equations (GEEs).
}
\examples{
data(musdata)
coords<- musdata[,4:5]

xgee<-GEE(musculus ~ pollution + exposure, "poisson", musdata,
      coord=coords, corstr="fixed", plot=TRUE,scale.fix=FALSE)

mwrm<-WRM(musculus ~ pollution + exposure, "poisson", musdata,
coord=coords, level=1, plot=TRUE)

mod.select.gee<-step.spind(xgee,musdata)
mod.select.wrm<-step.spind(mwrm,musdata)

}
\seealso{
\code{\link{aic.calc}},\code{\link{qic.calc}}
}
\author{
Sam Levin
}
