% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_spind.R
\name{step.spind}
\alias{step.spind}
\title{Stepwise AIC, AICc, and QIC for GEEs and WRMs}
\usage{
step.spind(object, data, steps = NULL, trace = TRUE, AICc = FALSE)
}
\arguments{
\item{object}{A model of class \code{WRM} or \code{GEE}.}

\item{data}{The data used to fit that model.}

\item{steps}{Number of iterations the procedure should
go through before concluding. The default is to use the number of
variables as the number of iterations.}

\item{trace}{Should R print progress updates to the console? Default
is \code{TRUE}.}

\item{AICc}{Logical. In the case of model selection with \code{WRM}s,
should AICc be used to determine which model is best rather than AIC?
This argument is ignored for \code{GEE}s. Default is \code{FALSE}.}
}
\value{
A data frame containing the results of the final iteration
of the stepwise model selection process. For \code{WRM}'s, the columns
returned are
\itemize{
  \item\code{Deleted.Vars} Variables deleted from the previous iteration.
  \item\code{LogLik} Log-likelihood of the model.
  \item\code{AIC} AIC score for the model.
  \item\code{AICc} AICc score for the model.
}

For \code{GEE}s:
\itemize{
  \item\code{Deleted.Vars} Variables deleted from the previous iteration.
  \item\code{QIC} Quasi-information criterion of the model.
  \item\code{Quasi.Lik} Quasi-likelihood of the model.
}
}
\description{
Stepwise AIC, AICc, and QIC for GEEs and WRMs
}
\details{
This function performs stepwise variable elimination
for model comparison. Each iteration will try to find the best
combination of variables for a given number of variables and
scope based on AIC, AICc, or QIC, and then use that as the base model
for the next iteration until there are no more variables to eliminate.
Alternatively, it will terminate when reducing the number of variables
while respecting the model heirarchy no longer produces lower
information criterion values.
}
\note{
Currently, the function only supports backwards model selection
(i.e. one must start with a full model and subtract variables).
Forward selection and both options may be added later.
}
\references{
Hardin, J.W. & Hilbe, J.M. (2003) Generalized Estimating Equations. Chapman and Hall, New York.
}
\seealso{
\code{\link{qic.calc}}, \code{\link{aic.calc}}, \code{\link[stats]{drop1}}
\code{\link[stats]{step}}, \code{\link[MASS]{stepAIC}}
}
\author{
Sam Levin
}
