% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_WRM.R
\name{summary.WRM}
\alias{summary.WRM}
\title{summary.WRM}
\usage{
\method{summary}{WRM}(object, ...)
}
\arguments{
\item{object}{A model of class \code{WRM}}

\item{...}{Other arguments passed to \code{summary}}
}
\value{
Prints results of model
}
\description{
Provides \code{summary} methods for WRM model objects
}
\author{
Gudrun Carl, Sam Levin
}
