% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_GEE.R
\name{summary_gee}
\alias{summary_gee}
\title{Summarize the output from\code{GEE}}
\usage{
summary_gee(model, printAutoCorMat = FALSE)
}
\arguments{
\item{model}{An object of class \code{gee}}

\item{printAutoCorMat}{A logical indicating whether to print the
autocorrelation matrix}
}
\value{
Prints model details, parameter estimates, and autocorrelation values
for the first 10 distance bins. Additionally, prints the autocorrelation matrix
if \code{printAutoCorMat} = TRUE
}
\description{
Returns summary of GEE parameter estimates and autocorrelations
of the residuals.
}
