% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_GEE.R
\name{summary_gee}
\alias{summary_gee}
\title{Summarize the output from\code{GEE}}
\usage{
summary_gee(model)
}
\arguments{
\item{model}{an object of class \code{gee}}
}
\description{
Returns summary of GEE parameter estimates and autocorrelations
of the residuals.
}
