% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_dep.R
\name{th.dep}
\alias{th.dep}
\title{Kappa and confusion matrix for species distribution models}
\usage{
th.dep(data, coord, thresh = 0.5, spatial = TRUE)
}
\arguments{
\item{data}{A data frame or matrix with two columns. The first column
should contain actual presence/absence data (binary, 0 or 1) and the
second column should contain model predictions of probability of
occurence (numeric, between 0 and 1).}

\item{coord}{A data frame or matrix with two columns containing x,y
coordinates for each actual and predicted value. Coordinates must be
integer and consecutively numbered.}

\item{thresh}{A cutoff value for classifying predictions as modeled
presences or modeled absences. Default is 0.5.}

\item{spatial}{A logical indicating whether spatially corrected indices
(rather than classical indices) should be computed.}
}
\value{
A list with the following components:
\describe{
 \item{\code{kappa}}{Kappa statistic}
 \item{\code{cm}}{Confusion matrix}
 \item{\code{sensitivity}}{Sensitivity}
 \item{\code{specificity}}{Specificity}
 \item{\code{actuals}}{Actual occurence data or adjusted actual occurence data}
 \item{\code{splitlevel.pred}}{Level splitting of predicted values}
 \item{\code{splitlevel.act}}{Level splitting of actuals or adjusted actuals}
 \item{\code{splitposition.pred}}{Position splitting of predicted values}
 \item{\code{splitposition.act}}{Position splitting of actuals or adjusted actuals}
}
}
\description{
Calculates threshold-dependent metrics for an observational
data set and model predictions. It is a 2-dimensional analysis that takes
the grid structure of the data into account. Grid cells are assumed to be
square
}
\examples{
data(hook)
data<-hook[,1:2]
coord<-hook[,3:4]
si1<-th.dep(data,coord,spatial=TRUE)
si1$kappa
si1$cm

}
\references{
Carl G, Kuhn I (2016)
Spind: a package for computing spatially corrected accuracy measures.
Ecography. DOI: 10.1111/ecog.02593
}
\author{
Gudrun Carl
}
