% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale.R
\name{upscale}
\alias{upscale}
\title{Upscaling of smooth components}
\usage{
upscale(f, x, y, wavelet = "haar", wtrafo = "dwt", pad = mean(f))
}
\arguments{
\item{f}{a vector}

\item{x}{corresponding x-coordinates which have to be integer}

\item{y}{corresponding y-coordinates which have to be integer}

\item{wavelet}{type of wavelet: "haar" or "d4" or "la8"}

\item{wtrafo}{type of wavelet transform: "dwt" or "modwt"}

\item{pad}{a numeric value for padding into a bigger square.
Mostly the mean of f values is useful.
Default is set to mean(f).}
}
\description{
The analysis is based a wavelet multiresolution analysis
using only smooth wavelet components.
It is a 2D analysis taking the grid structure of datasets into
account, i.e. the analysis provides scale-specific
results for data sampled on a contiguous geographical area. The
dataset is assumed to be regular gridded and the grid cells are
assumed to be square.
The scale-dependent results are graphically displayed.
}
