% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavevar.R
\name{wavevar}
\alias{wavevar}
\title{Wavelet variance analysis}
\usage{
wavevar(f, x, y, wavelet = "haar", wtrafo = "dwt")
}
\arguments{
\item{f}{a vector}

\item{x}{corresponding x-coordinates which have to be integer}

\item{y}{corresponding y-coordinates which have to be integer}

\item{wavelet}{type of wavelet: "haar" or "d4" or "la8"}

\item{wtrafo}{type of wavelet transform: "dwt" or "modwt"}
}
\value{
Wavelet variance
}
\description{
The analysis is based a wavelet multiresolution analysis.
It is a 2D analysis taking the grid structure of datasets into
account, i.e. the analysis provides scale-specific
results for data sampled on a contiguous geographical area. The
dataset is assumed to be regular gridded and the grid cells are
assumed to be square.
}
\seealso{
\pkg{waveslim},\code{\link{WRM}}
}
