% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavevar.R
\name{wavevar}
\alias{wavevar}
\title{Wavelet variance analysis}
\usage{
wavevar(f, x, y, wavelet = "haar", wtrafo = "dwt")
}
\arguments{
\item{f}{A vector}

\item{x}{Corresponding x-coordinates which have to be integer}

\item{y}{Corresponding y-coordinates which have to be integer}

\item{wavelet}{Name of wavelet family. \code{haar}, \code{d4}, and \code{la8}.
are possible. \code{haar} is the default.}

\item{wtrafo}{Type of wavelet transform. Either \code{dwt} or \code{modwt}.
\code{dwt} is the default.}
}
\value{
Wavelet variance for \code{f}.
}
\description{
Calculates the wavelet variance based on a wavelet
multiresolution analysis.
}
\seealso{
\pkg{waveslim}, \code{\link{WRM}}, \code{\link{covar.plot}}
}
\author{
Gudrun Carl
}
